pyproject.toml
[project]
name = "dotsales"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crispy-tailwind>=1.0.3",
    "django-cors-headers>=4.8.0",
    "django-cotton>=2.1.3",
    "django-htmx>=1.24.0",
    "django-ninja>=1.4.3",
    "django-ninja-extra>=0.30.1",
    "django-ninja-jwt>=5.3.7",
    "django-storages[azure]>=1.14.6",
    "django-stubs>=5.2.5",
    "django-unicorn>=0.62.0",
    "django-widget-tweaks>=1.5.0",
    "djangosetupheist>=0.6.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "resend>=2.13.1",
]
